+++ src/interface/rest-controller/src/route/mod.rs
@@ -1,0 +1,1 @@
+pub mod user;
